agentic-framework-main/pyproject.toml
[project]
name = "agentic-framework"
version = "1.0.0"
description = "Enterprise Agentic Framework for Multi-Agent Systems - LLM-agnostic orchestration platform"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Agentic Framework Contributors", email = "dev@agentic-framework.org"}
]
keywords = ["ai", "agents", "llm", "orchestration", "multi-agent", "workflow", "anthropic", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "pymilvus>=2.3.0",
    "chromadb>=0.4.18",
    "minio>=7.2.0",
    "celery>=5.3.0",
    "sentence-transformers>=2.2.2",
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "prometheus-client>=0.19.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "anyio>=4.0.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0.0",
    "python-socketio>=5.8.0",
    "websockets>=11.0.0",
    "passlib[bcrypt]>=1.7.4",
    # CLI Tool (Kautilya) dependencies
    "click>=8.1.7",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "questionary>=2.0.1",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "ddgs>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/agentic-framework"
Documentation = "https://github.com/your-org/agentic-framework/blob/main/docs"
Repository = "https://github.com/your-org/agentic-framework"
Issues = "https://github.com/your-org/agentic-framework/issues"
Changelog = "https://github.com/your-org/agentic-framework/blob/main/CHANGELOG.md"

[project.scripts]
kautilya = "tools.kautilya.kautilya.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests", "*/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["adapters*", "subagent_manager*", "orchestrator*", "memory_service*", "code_exec*", "mcp_gateway*", "tools*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
